pyproject.toml
[project]
name = "diversity-django-london"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "flask>=3.0.3",
    "openai>=1.21.2",
    "pandas>=2.2.2",
    "openpyxl>=3.1.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
